First attempt to fix gif animation rendering on GitHub pages
In the README file, I can insert the gif with the usual markdown command and relative local path. This will be rendered on the GitHub repo. But this fails for the markdown files used to generate the pages on the static web site. While locally it works, once pushed online, it fails. I tried then to use the raw URL of the gif from the repo.